♻️ refactor(cli): extract run-plan vertical slice orchestration
